ipaddpool/requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.1.0
Flask-Cors==3.0.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
six==1.12.0
SQLAlchemy==1.3.5
Werkzeug==0.15.4
